+++ truffle/src/com.oracle.truffle.api.bytecode/src/com/oracle/truffle/api/bytecode/BytecodeAccessor.java
@@ -52,7 +52,6 @@ final class BytecodeAccessor extends Accessor {
     static final EngineSupport ENGINE = ACCESSOR.engineSupport();
     static final MemorySupport MEMORY = ACCESSOR.memorySupport();
     static final RuntimeSupport RUNTIME = ACCESSOR.runtimeSupport();
-    static final EngineSupport ENGINE = ACCESSOR.engineSupport();
     static final LanguageSupport LANGUAGE = ACCESSOR.languageSupport();
 
     private BytecodeAccessor() {