+++ truffle/src/com.oracle.truffle.api.debug/src/com/oracle/truffle/api/debug/DebugValue.java
@@ -57,6 +57,7 @@ import java.util.function.Consumer;
 
 import com.oracle.truffle.api.CompilerDirectives;
 import com.oracle.truffle.api.TruffleLanguage;
+import com.oracle.truffle.api.instrumentation.TruffleInstrument;
 import com.oracle.truffle.api.interop.InteropLibrary;
 import com.oracle.truffle.api.interop.InvalidArrayIndexException;
 import com.oracle.truffle.api.interop.StopIterationException;
@@ -1873,24 +1874,9 @@ public abstract class DebugValue {
         InteropLibrary lib = InteropLibrary.getFactory().getUncached(obj);
         if (lib.hasLanguageId(obj)) {
             try {
-                return getSession().getDebugger().getEnv().getLanguageInfo(lib.getLanguageId(obj));
-            } catch (UnsupportedMessageException e) {
-                CompilerDirectives.transferToInterpreter();
-                throw new AssertionError(e);
-            }
-        }
-        return getOriginalLanguageDeprecated(lib, obj);
-    }
-
-    /**
-     * GR-69615: Remove deprecated InteropLibrary#hasLanguage and InteropLibrary#getLanguage
-     * messages.
-     */
-    @SuppressWarnings("deprecation")
-    private LanguageInfo getOriginalLanguageDeprecated(InteropLibrary lib, Object obj) {
-        if (lib.hasLanguageId(obj)) {
-            try {
-                return getSession().getDebugger().getEnv().getLanguageInfo(lib.getLanguageId(obj));
+                String languageId = lib.getLanguageId(obj);
+                TruffleInstrument.Env env = getSession().getDebugger().getEnv();
+                return "host".equals(languageId) ? env.getHostLanguage() : env.getLanguages().get(languageId);
             } catch (UnsupportedMessageException e) {
                 CompilerDirectives.transferToInterpreter();
                 throw new AssertionError(e);